django_trips/requirements.txt
setuptools
# Stay on the latest LTS release of Django
Django<2.3
django-cors-headers
django-config-models
django_extensions
djangorestframework
mysqlclient
factory_boy
pytest
